Move cached collections for settings
